% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepocessing.R
\name{applyCountThreshold}
\alias{applyCountThreshold}
\title{Filter genes}
\usage{
applyCountThreshold(data, metadata, threshold = 32)
}
\arguments{
\item{data}{data.frame. Raw count data to filter.}

\item{metadata}{data.frame. Metadata corresponding to the \code{data}.}

\item{threshold}{integer. Threshold below which gene are going to be removed}
}
\value{
A data.frame which is a subset of \code{data} for which genes that have all
counts below the threshold were removed
}
\description{
Filter out genes for which no experiment pass the given threshold.
}
\examples{
attach(MEF_dataset)
filtered <- applyCountThreshold(full_counts, metadata)
}
