% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepocessing.R
\name{checkMetadata}
\alias{checkMetadata}
\title{Check potential error in the metadata file}
\usage{
checkMetadata(metadata, counts, type = "required")
}
\arguments{
\item{metadata}{data.frame. Metadata data frame to check for consistency.
The following columns are expected:
\tabular{ll}{
       Orgamism \tab organism used for the experiment \cr
       Genotype: \tab genotype of the organism used for the experiment \cr
       CellType: \tab cell type used for the experiment \cr
       Conditionning: \tab conditioning used for the experiment \cr
       ExperimentNumber: \tab number of a single ActinomycinD timecourse \cr
       ExperimentName: \tab name given of a single ActinomycinD timecourse \cr
       ExperimentReplicate: \tab replicate number for the experiment \cr
       ActDTime: \tab time after ActinomycinD treatment in min \cr
       Stimulus: \tab stimulus used for the experiment \cr
       StimulusTime: \tab time between addition of the stimulus and ActinomyinD in min\cr
       SampleName: \tab name of the sample. Must correspond to colnames of
       full_counts \cr
       QC: \tab by default "Included", if for some reason the sample needs to be
       removed then can be changed to "QC failed" \cr
       GEOid: \tab if data are deposited in GEO, then GEOid of the sample \cr
}}

\item{counts}{data.frame. The counts data  related to the \code{medata}.}

\item{type}{character. Either 'all' or 'required'. Against which columns to
check consistency, all of them or only the required ones.}
}
\description{
Check metadata file for obvious discrepancy such SampleName column not
matching column names of the counts, not having required columns ...
}
\examples{
attach(MEF_dataset)
checkMetadata(metadata, full_counts, type='all')
}
