% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysing.R
\name{convertSlopeToHL}
\alias{convertSlopeToHL}
\title{Convert slope result to HL results}
\usage{
convertSlopeToHL(result)
}
\arguments{
\item{result}{named list of data.frame. Each element of the list correspond
to the slope result of one experiment. This is obtained from \code{\link{inferDecay}}.}
}
\value{
A list containing half-life results for all genes and all experiments.
Each element corresponds to the result of a single experiment with the following
information:
\tabular{ll}{
  HL \tab estimated half-life in minute \cr
  CI_X_LB \tab lower bound of the X\% confidence interval of the half-life \cr
  CI_X_UB \tab upper bound of the X\% confidence interval of the half-life \cr
  Adj Rsquare \tab adjusted R square of the regression \cr
}
}
\description{
Convert slope results obtained with \code{\link{inferDecay}} to half-lives
results
}
\details{
This function will try all the possible linear regression starting at any
point within the first hour of ActinomycinD treatment and for which the
counts are within the max counts within the first hour and the max within
the first hour -1/4 of the max decay rate per hour. Then it will try all
possible linear regressions allowing one intermediary point to be excluded
and with at least 3 remaining points. It will select the fit with  the best
adjusted R^2 and a negative slope as result.
}
