% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysing.R
\name{decay_with_replicates_1exp}
\alias{decay_with_replicates_1exp}
\title{Get fit result for a specific gene and experiment}
\usage{
decay_with_replicates_1exp(
  counts,
  time,
  replicates = rep(1, length(counts)),
  qc,
  threshold
)
}
\arguments{
\item{counts}{numeric vector. log2 normalised counts for a specific experiment.}

\item{time}{numeric vector. Corresponding time after ActinomycinD treatment
for a specific experiment.}

\item{replicates}{integer vector. Corresponding replicate number as given by
\code{metadata$ExperimentReplicate}.}

\item{qc}{character vector. Correspondinc QC status as given by
\code{metadata$QC}.}

\item{threshold}{numeric. log2 threshold for which a replicate with all counts
below it will not be considered.}
}
\value{
A vector containing fit results with the following information:
\tabular{ll}{
  ReplicatesNumber \tab replicates number from \code{metadata$ExperimentNumber}
  used, if multiple replicates were used they are separated by a comma \cr
  Start \tab indice of the starting point for the regression , if multiple
  replicates were used they are separated by a comma \cr
  End \tab indice of the point that was skipped for the regression , if multiple
  replicates were used they are separated by a comma \cr
  Intercept \tab intercept result for the regression , if multiple
  replicates were used they are separated by a comma \cr
  Slope \tab slope result of the regression \cr
  Adj Rsquare \tab adjusted R square of the regression \cr
  DF \tab degrees of freedoms for the regression \cr
  CI_X_LB \tab lower bound of the X\% confidence interval of the slope \cr
  CI_X_UB \tab upper bound of the X\% confidence interval of the slope \cr
}
}
\description{
Get fit result for a specific gene including replicates information.
All possible fit will be tried and the one with the best adjusted R^2
will be selected.
}
\details{
This function will try all the possible linear regression starting at any
point within the first hour of ActinomycinD treatment and for which the
counts are within the max counts within the first hour and the max within
the first hour -1/4 of the max decay rate per hour. Then it will try all
possible linear regressions allowing one intermediary point to be excluded
and with at least 3 remaining points. It will select the fit with the best
adjusted R^2 as result.
}
\keyword{internal}
