% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{evalLayer}
\alias{evalLayer}
\title{Evaluate ggplot layer added based on control list}
\usage{
evalLayer(p, layer, args)
}
\arguments{
\item{p}{ggplot2 object. Current plot to add a layer to.}

\item{layer}{character. gpplot2 function to use.}

\item{args}{named list. Arguments names and their associated value of the
\code{layer} function.}
}
\value{
A gggplot2 object.
}
\description{
Evaluate ggplot layer added based on control list
}
\keyword{internal}
