% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepocessing.R
\name{extractSpikeIn}
\alias{extractSpikeIn}
\title{Separate spike-ins}
\usage{
extractSpikeIn(data, spikein = "ERCC")
}
\arguments{
\item{data}{data.frame. filtered count data as the result of
\code{\link{applyCountThreshold}}.}

\item{spikein}{character. Pattern matching in the spike-ins names.}
}
\value{
A list of 2 data.frames. The first on is the a subset of \code{data}
with only the spike-ins extracted, the second one corresponds to the genes counts
without the spike-ins
}
\description{
Separate spike-ins from gene data
}
\examples{
attach(MEF_dataset)
tmp <- extractSpikeIn(filtered, spikein = "ERCC")
spikeins <- tmp[[1]]
counts <- tmp[[2]]
detach(MEF_dataset)
}
