% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysing.R
\name{inferDecay}
\alias{inferDecay}
\title{Infer slope from normalised data}
\usage{
inferDecay(
  norm_counts,
  metadata,
  threshold = 32,
  conf_levels = c(0.95, 0.75, 0.5)
)
}
\arguments{
\item{norm_counts}{numeric matrix or data.frame. Normalised data
as given by the second element of the result of \code{\link{SpikeInsNormalisation}}.}

\item{metadata}{data.frame. Metadata of the data. Must have specific column
names see \code{\link{checkMetadata}} function.}

\item{threshold}{numeric. If an experiment as all normalised counts below
this threshold no fit are going to be perfomed.}

\item{conf_levels}{numeric vector. Value of the confindence intervals wanted.
Values must be between 0 and 1. If it is a vector, it will calculate the confidence
intervals for all the values given.}
}
\value{
A list which elements corresponds to the result of one experiment. \cr
Each element is a data.frame with the following columns:
\tabular{ll}{
  ReplicatesNumber \tab replicates number from \code{metadata$ExperimentNumber}
  used, if multiple replicates were used they are separated by a comma \cr
  Start \tab indice of the starting point for the regression , if multiple
  replicates were used they are separated by a comma \cr
  End \tab indice of the point that was skipped for the regression , if multiple
  replicates were used they are separated by a comma \cr
  Intercept \tab intercept result for the regression , if multiple
  replicates were used they are separated by a comma \cr
  Slope \tab slope result of the regression \cr
  Adj Rsquare \tab adjusted R^2 of the regression \cr
  DF \tab degrees of freedoms for the regression \cr
  CI_X_LB \tab lower bound of the X\% confidence interval of the slope \cr
  CI_X_UB \tab upper bound of the X\% confidence interval of the slope \cr
}
}
\description{
Infer slope and its confidence interval from the ActinomycinD normalised data.
Depending on the number of genes and experiment it can take some time.
Most use to plot the result of single genes.
}
\details{
This function will try all the possible linear regression starting at any
point within the first hour of ActinomycinD treatment and for which the
counts are within the max counts within the first hour and the max within
the first hour -1/4 of the max decay rate per hour. Then it will try all
possible linear regressions allowing one intermediary point to be excluded
and with at least 3 remaining points. It will select the fit with the best
adjusted R^2 as result.
}
\examples{
attach(MEF_dataset)
decay <- inferDecay(norm_counts[1:100,], metadata)
detach(MEF_dataset)
}
