% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotLibrarySize}
\alias{plotLibrarySize}
\title{Plot bargraph of library size before and after normalisation}
\usage{
plotLibrarySize(before, after, metadata, control = NULL)
}
\arguments{
\item{before}{numeric matrix or data.frame. Gene counts matrix before
normalisation.}

\item{after}{numeric matrix or data.frame. Gene counts matrix after
normalisation.}

\item{metadata}{data.frame. Metadata of the data. Must have specific column
names see \code{\link{checkMetadata}} function.}

\item{control}{list. Further control ggplot2 output. This must be a named
list with the names corresponding to ggplot2 function and the associated
value is also a named list of the ggplot2 function argument name and its
associated value.}
}
\description{
This function plots observed library size before normalisation and effective
normalised library size.
It assumes that before and after colnames are found in the column
SampleName of the metadata
}
\examples{
attach(MEF_dataset)
plotLibrarySize(before = counts, after = norm_counts, metadata = metadata)

#' # Add some control over the ggplot2 output
plotLibrarySize(before = counts, after = norm_counts, metadata = metadata,
                control = list(labs=list(y="Effective Library size")))
detach(MEF_dataset)
}
