% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA gene counts before and after normalisation}
\usage{
plotPCA(before, after, metadata, dim1 = 1, dim2 = 2, control = NULL)
}
\arguments{
\item{before}{numeric matrix or data.frame. Ggene counts matrix before
normalisation.}

\item{after}{numeric matrix or data.frame. Gene counts matrix after
normalisation.}

\item{metadata}{data.frame. Metadata of the data. Must have specific column
names see \code{\link{checkMetadata}} function.}

\item{dim1}{integer. Principal component to plot in x axis e.g. 1 will plot
the first principal component as x axis.}

\item{dim2}{integer. Principal component to plot in y axis e.g. 2 will plot
the second principal component as y axis.}

\item{control}{list. Further control ggplot2 output. This must be a named
list with the names corresponding to ggplot2 function and the associated
value is also a named list of the ggplot2 function argument name and its
associated value.}
}
\description{
This function plots PCA of gene counts before and after normalisation.
It assumes that before and after colnames are found in the column
SampleName of the metadata
}
\examples{
attach(MEF_dataset)
plotPCA(before = counts, after = norm_counts, metadata = metadata)
}
