% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plotSpikeIns}
\alias{plotSpikeIns}
\title{Plot boxplots of spike-ins counts before and after normalisation}
\usage{
plotSpikeIns(before, after, metadata, control = NULL)
}
\arguments{
\item{before}{numeric matrix or data.frame. Spike-ins counts matrix before
normalisation.}

\item{after}{numeric matrix or data.frame. Spike-ins counts matrix after
normalisation.}

\item{metadata}{data.frame. Metadata of the data. Must have specific column
names see \code{\link{checkMetadata}} function.}

\item{control}{list. Further control ggplot2 output. This must be a named
list with the names corresponding to ggplot2 function and the associated
value is also a named list of the ggplot2 function argument name and its
associated value.}
}
\description{
This function plots spikeins counts before and after normalisation.
It assumes that before and after colnames are found in the column
SampleName of the metadata
}
\examples{
attach(MEF_dataset)
plotSpikeIns(before = spikeins, after = norm_spikeins, metadata = metadata)

# Add some control over the ggplot2 output
plotSpikeIns(before = spikeins, after = norm_spikeins, metadata = metadata,
             control = list(labs=list(title= "Spike-ins counts before vs after normalisation")))
detach(MEF_dataset)
}
