% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysing.R
\name{regression_function}
\alias{regression_function}
\title{Get fit result for a specific regression}
\usage{
regression_function(counts, time, indices, start, end, skip, rep_order)
}
\arguments{
\item{counts}{list of numeric vectors. List of log2 normalised counts for a
specific experiment with each element correpsonding to one replicate.}

\item{time}{list of numeric vectors. List of corresponding time after
ActinomycinD treatment for a specific experiment, with each element correpsonding to one
replicate.}

\item{indices}{list of integers vectors. List of indices of useable counts,
with each element correpsonding to one replicate.}

\item{start}{integers vector. Index of the starting point for the regression,
with each element correpsonding to one replicate.}

\item{end}{integers vector. Index of the ending point for the regression,
with each element correpsonding to one replicate.}

\item{skip}{integers vector. Index of skipped point for the regression if any,
with each element correpsonding to one replicate.}

\item{rep_order}{integers vector. Replicate number giving the order of the
previous parameters}
}
\value{
A vector containing fit results with the following information:
\tabular{ll}{
  ReplicatesNumber \tab replicates number from \code{metadata$ExperimentNumber}
  used, if multiple replicates were used they are separated by a comma \cr
  Start \tab indice of the starting point for the regression , if multiple
  replicates were used they are separated by a comma \cr
  End \tab indice of the point that was skipped for the regression , if multiple
  replicates were used they are separated by a comma \cr
  Intercept \tab intercept result for the regression , if multiple
  replicates were used they are separated by a comma \cr
  Slope \tab slope result of the regression \cr
  Adj Rsquare \tab adjusted R square of the regression \cr
  DF \tab degrees of freedoms for the regression \cr
  CI_X_LB \tab lower bound of the X\% confidence interval of the slope \cr
  CI_X_UB \tab upper bound of the X\% confidence interval of the slope \cr
}
}
\description{
Get fit result for a specific regression including replicates information
}
\keyword{internal}
