% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepocessing.R
\name{SpikeInsNormalisation}
\alias{SpikeInsNormalisation}
\title{Normalisation based on spike-ins}
\usage{
SpikeInsNormalisation(
  spikein_counts,
  genes_counts,
  gene_infos,
  metadata,
  method = "RLE",
  p = 0.75
)
}
\arguments{
\item{spikein_counts}{data.frame. Extracted spike-ins count data as the first
element of the result from \code{\link{extractSpikeIn}}.}

\item{genes_counts}{data.frame. Extracted gene count data as the second
element of the result from \code{\link{extractSpikeIn}}.}

\item{gene_infos}{data.frame. Information on the genes. It must include at
least a Geneid column which match the row names of the \code{gene_counts}.}

\item{metadata}{data.frame. Metadata corresponding to the \code{gene_counts}.}

\item{method}{character. Method to get normalisation factors based on the
spike-ins. Must be one method recognized by \code{\link[edgeR]{calcNormFactors}}.
By default use "RLE" method.}

\item{p}{numeric. Percentile (between 0 and 1) of the counts that is aligned
when method="upperquartile".}
}
\value{
A list of 2 data.frames. The first on is the normalised version of
\code{spikein_counts}, the second one corresponds to the normalised version of
\code{genes_counts}.
}
\description{
Normalised counts and spike-ins based on spike-ins counts
}
\examples{
attach(MEF_dataset)
tmp <- SpikeInsNormalisation(spikeins, counts, gene_infos, metadata)
norm_spikeins <- tmp[[1]]
norm_counts <- tmp[[2]]
detach(MEF_dataset)
}
